Refactor to use template method
